3.0

+++ run.ps1
@@ -8,9 +8,20 @@ git clone $repository  $workspace
 
 # Add devcontainer, Jenkinsfile and post-commit files
 
-mv ./.devcontainer ./${workspace}/
-mv ./post-commit ./${workspace}/.git/hooks/
+if(![System.IO.File]::Exists(./.devcontainer)){
+    throw "Clone already done."
+}
+else {
+    mv ./.devcontainer ./${workspace}/
+    mv ./post-commit ./${workspace}/.git/hooks/
+}
+
 
 docker-compose build
 
 docker-compose up -d
+
+
+
+
+
